Add "Mail Page" action which mails the current webpage.
It creates a new mail in Sparrow.app, fills the subject with the page
title and places the URL of the page on the body.
It leaves the caret on the recipient field so you can take advantage
of  the email client's auto-completion ability.
It currently only works with google chrome but could be extended to
work with other browsers and email clients.